Consolidated fit function into singlet, removed fit.py
